Fixed broken route resolution
Originally all routes under the slash domain would match with the flags
route. This meant that routes like '/login' would default to '/flags' if
the route didn't exist. This change now removes that functionality and
will properly display a 404 (The default NextJS one).

Notes:
Routing is a little odd in pages router of NextJS. Although the dir was
[...flags], it was using the catch-all segments logic of [[...flags]]
instead. I remedied this by creating a concrete flags route and then
adding in the slug routes for projectId and flagId. This also broke the
FlagNav as it was hardcoded to take the user to /flags/projectId/flagId.
It would result in the user going to /flags/flags/porjectId/flagId. This
kind of routing is worrisome since it is not typed, and can cause ripple
errors like this one. In the future, I should use some sort of redux-
like route handling to ensure that all links go to the right places
without the need for hardcoding them within components.

I also need to create a default 404 page as currently the NextJS one is
used and it does not fit the style of the app.

I also will need to rewrite how the ReactFlow portion handles state
changes as currently, whenever a route changes, there are 2-3 rerenders
that occur. I have a feeling it is a mix of the useEffects and some
ReactFlow issue. I might just create my own graph library, but that
seems like a last resort.