Supporting multiple method types on a route
Description:
- Changed route resolver to now check for routes by their method type
- Changed route resolver to now check for schemas by their method type

Test Plan:
Checked that all tests pass and postman still returns valid responses

Issue:
https://github.com/IzzyBeraja/Flags/issues/22

Notes:
Previously, every route had a route which would be used in conjunction
with the method constant in every route file. This made it such that
only one method could be used per file. This was fine and worked with
the schema validator. I wanted to support multiple methods so I debated
creating a methods folder for each of the routes and use the name of the
file to determine the method. I think that this would result in too many
files and I assume there is repeat logic between routes with different
methods anyways. To ensure that each route could have custom validators,
I decided on a new naming convention.

Each route needs to define a method as a const export. If that method
has some sort of validation, it should have an exported schema with the
name <method>RequestSchema. This can also be done for the response as
well, but currently I am not doing anything with it. I think in the
future, I will attach my docs to it.

I may create additional fields in the future to determine if a route
should be active or inactive, or if they should be visible to public or
devs only. That however is for a future time.