Fixed duplicate route edge case
Description:
- Removed support for index -> / route
- Fixed duplicate route checkes

Test Plan:
- Ensure that all routes still resolve as before

Notes:
I realize that there is an edge case that I didn't account for that can
result in duplicate routes. Originally, I supported the use of index as
a name for a route so that the / directory of that route could be made.
Although this seemed rather benign, I realize now that there is a way to
create duplicate routes this way.

Example: /api/users/@me AND /api/users/@me/

If you create a file called @me.routes.ts within the users directory and
create an index.routes.ts within the @me directory, you get two almost
identical routes. With some testing, I found that my implementation
would only run the route ending in /. I recognize now that using index
can be problematic so I have removed that functionality. This should
prevent cases of duplicate routes in the future (unless some sort of
linking is possible, but I don't do that...)

I also found that after switching to the new Set implementation to
expose route metadata, I forgot that object comparison is shallow and
based on the reference so routes could never truly be equal to each
other. I have since switched to a Map and use a combination of method
and route path as the key. This should help catch duplicates. I also
thought about removing trailing /'s, but I need to think about it a
little more before I commit to that.