Added PATCH route for @me and password route
Description:
- Added PATCH route for @me so fields can be updated
- Added PUT route at @me/password so password can be changed

Test Plan:
- Added tests for PATCH route
- Added tests for PUT route

Issue:
https://github.com/IzzyBeraja/Flags/issues/23

Notes:
Added a way for the current user to modify their information. Currently
this is only available for the name field, but in the future this will
also be possible for email and other fields.

I chose to put the password modification in its own route because it has
some additional details that are required to properly update the pass.
I'm also doing this because I think I will be moving the password to its
own table in the DB so that I can query the Users table without having
to ensure that in every place I don't accidentally send the password
field. That does require some additional logic so I haven't done that
migration just yet.

Additionally I noticed that I used bcrypt in a couple locations so I
moved the hashing and comparing functions to their own file within utils
so that I wouldn't have to repeat storing the salt rounds variable. I
can also create an env variable for it, but this seems easier anyways if
I decided to switch algorithms in the future.

I also am noticing that my unit tests are rather involved. I don't think
that for CRUD operations like these there is really much else I can do
here to make them easier. They might be simpler if I use a message
queue in the future but that will only offset the testing to the DB
service.