Adding create route for killswitch
Description:
- Creates a new killswitch given a name and some fields

Notes:
There are some issues with this, namely error handling. I have to ensure
that at each stage, I am properly checking that each of the queries does
not fail before continuing. Although I can currently do this, Prisma
does not provide adequate errors relating to the constraints that have
failed which is making it harder to return a tailored message to the
client. Will look into this further in the future.