Updating register account using slonik
Description:
- api/auth/register now creates an account rather than a user
- route now uses slonik rather than prisma

Test Plan:
- There are a few errors currently, will fix them all in a future commit

Notes:
I've been debating between using prisma or not for quite a while. For a
moment, I thought that I would end up using it, but ran into an issue.
I plan on storing flags within a diagram table that requires a CTE to
query the trees. This is a behavior that is not supported in Prisma and
rather simple to implement in SQL directly. As such, I've decided to
move away from prisma in stages in favor of slonik (an sql quey builder)
as it does not inhibit me.

There are some issues to address here. Currently slonik requires the use
of Zod to handle typing. It uses it as a form of verification that the
types returned from SQL are the same as they say they are. I'm debating
whether or not to use Zod as if I do, I might as well move away form AJV
and use Zod instead. This seems like a rather large change so I'd like
to delay it if possible.

Another thing to mention is that currently, I will be using a new
database management system and it will be the last change I make here. I
was trying out prisma with cockroachdb and found that the error handling
was mildly better, however I've thought it over and plan to use
cockroachdb instead of planetscale. At this point, I just want to have a
non-changing decision and cockroachdb just happens to be what I've been
using for the past couple days. I will start to remove my AWS db and
planetscale in the near future. Other reasons for this change involve
foreign key support, easier constraint management, and that cockroachdb
handles online schema changes in a way that I think is comparable to
planetscale. I don't plan on revisiting this for a long while as I've
spent far too long figuring out what I should be using rather than
building.