Making Node Environment clearer
Description:
- Added assertions that environent variables are set
- Added Chalk
- Added clearer indications when in dev or prod mode

Notes:
I just spent a few hours trying to debug my code because session cookies
were not being sent to the client. I had no idea what was going on and
tried to load older versions of the code, but no luck! I found out that
my environment variables were wrong. I defined NODE_ENV twice and the
second one was production... I have made it a bit clearer which version
of the environment I am in and added some checks to ensure that env
variables are properly set before running the database.

In the future, I'd also like to set up logging for environment variable
issues as it will be hard to see what is going on with servers if they
simply shutdown gracefully without fully starting up. I'll come back to
this once I have established logging.