Updated all queries to Drizzle
Description:
- Created createUser
- Created getUser
- Created updateUser
- Created updateAccountCredentials
- Fixed session mislabling of userId to accountId

Issue:
[Migrate to Drizzle](https://github.com/IzzyBeraja/Flags/issues/34)

Notes:
Lots of changes here, but nothing too complicated. I used drizzle
transactions for the first time and they seem easy enough to use. I will
need to go back and properly do error handling as currently it's a bit
crazy.

As well, switch creation is not working since the session does not have
a proper userId. I can fix this when userId is added to the session. I
will need to add some checks to ensure that userId exists as well.