Makes an SQL request on startup
Description:
- Added a database request on startup

Notes:
The first request on my database was taking a while which I assume is
due to pool creation. It also is probably due in part to the database
server going cold. This should help with the first reuqest, but I will
need to make changes in the future.

I've also noticed that due to my cache not existing where my db does, my
connections are rather slow. On average ~80ms for a cache request and
~50ms for a database request. If they both lived in the same datacenter,
this could probably be dropped down to ~20ms total.