Add additional session data to cookie
Description:
- Cookie now supports additional information including:
  - accountId
  - userId
  - email
  - firstName
  - lastName
  - userAgent
  - ipAddress
  - loginDate
- Account login now properly adds values to cookie
- PATCH, POST routes on @me also update cookie

Issue:
[Add additional data within session cookie](https://github.com/IzzyBeraja/Flags/issues/33)

Notes:
Cookie stored within Redis cache now contains additional information
about the current user including what was mentioned in the description.
This should facilitate with sessions and logging.

Currently feeling the effects of the cache and the db not existing in
the same place. This is most notable on login where the cache gets
checked for the current session, then a db transaction happens which
requires two queries (one to check current password, next to get the
current information), and then finally updating the cache again. I
could make this easier by passing all db information in one shot, but
that still doesn't fix the two cache round trips. For now, this is okay
as login is expected to be slow, but I would like to investigate hosting
options sooner rather than later to get what could be ~100ms per request
in speed gains.

On another note, I would also like to validate the session in the future
using the ip adress and user agent fields using some sort of binding as
recommended by the [OWASP Cheet Sheet Series](https://cheatsheetseries.owasp.org/cheatsheets/Session_Management_Cheat_Sheet.html#binding-the-session-id-to-other-user-properties). When I include JWT
validation, I want to ensure that it is the same user. That might
require writing a custom session handler or simply another middleware
that checks that the request and session have the same data. The latter
would be easier to implement because I could send the JWT as part of the
body in the login response and add it during login. I will need to
figure out how to invalidate the token though which could be hard. I
will need to do some research.