Fixing the switch create route
Description:
- Updated switch route to properly store switches
- Updated some killswitch references to switch
- Fixed state not being stored
- Added unqiueness relationship between owned_by and name of switch

Issue:
[Switch create route](https://github.com/IzzyBeraja/Flags/issues/24)

Notes:
Realized that killswitches was still the name of the table in my
database to I updated it to switches. Only now, it's clicking that the
name switch is a reserved keyword in JS and almost all other languages
so I may need to come up with a better name.

The issue with the previous implementation is that switch is tied to the
userId and not the accountId. This means that if the account does not
have an associated user, it would fail. This was also an issue because
the userId was not being saved in the session storage and was passing
the accountId instead. This is now fixed.

I also added a uniqueness relationship between the user_id and the name
of the killswitch so that only the user can only have one switch with
that name. In the future when I add organizations and projects, I will
probably alter that constraint.

In the next diffs, I will be adding some indexes which is best practice
for foreign keys since they are not added automatically. I will also
add some collation rules so that case and accents do not matter when
determining the uniqueness of a field. This should make it impossible to
have two switches with similar names.