Adds a new route to get singular switch
Description:
- Adds a GET route which takes switchId as a param

Notes:
This works for now, but I would like to move toward codegen in the
future. I also don't know if params is a vector for attack so I will
need to do more research on it.

In the future, I will also add a route that simply returns the computed
states for each of the switches without the additional info.

I also need to figure out a name for switches. I may