Adds a new route to update switches
Description:
- Updated the name of the :id field to :switchId
- Added a new route that allows for updates of switches
- Requires that the userId is associated with the owner of the switch

Notes:
Currently to update the switch, the user needs to be logged in, and be
the owner of the switch. They can they update any of the fields
including name, description, and state.

I will need to do a few things in the future:
- Add an index on the name, owned_by, and state columns
- Add proper error handling for failed updates

I also need to ensure that my router does not allow for two routes that
share the same path with only the param having a different name
otherwise I will have issues with finding the correct route.