Fixed broken tests and other things
Description:
- Updated Jest in attempt to fix "verbatimModuleSyntax" (didn't work)
- Updated Jest.config.js -> Jest.config.ts and now supports tsconfig
- Updated Account type of ReqisterAccount (id -> accountId)
- Removed references of Prisma
- Updated type structure of routes to a new system
- Added new EmptyObject type due to empty interface problems

Test Plan:
- Ensure that all tests are passing (FINALLY)

Issue:
[Fix all broken tests](https://github.com/IzzyBeraja/Flags/issues/39)

Notes:
I updated Jest because there was an issue with verbatimModuleSyntax in
the tsconfig, but doing that didn't work. It said something along the
lines of an issue with commonJS which didn't make any sense to me. I do
however like that it now works with the typescript variant so I've kept
it.

I finally removed all references of Prisma in the tests. Since many of
them were actively running Prisma queries instead of the method I have
now of running a separate module that actually runs the query, they had
a lot of boilerplate code that was simple to remove. Drizzle has also
made working with these mocks easier and has resulted in less trouble
for now. I will still need to actually mock Drizzle in the future, but
things are fine for now.

I also updated the types for a lot of routes. There were some issues
with mocking due to having to encapsulate request and response types in
their express variants (Request<GetRequest>, Response<GetResponse>). I
instead opted to use the Parameters utility function on the handler
itself which has made typing significantly easier. As a result, I no
longer need to export the types within the route itself. I also added
the EmptyObject utility type because of issues with how Typescript
handles empty interface types. It was causing things like the json
responses to have weird types and requests could contain anything. This
should now fix that issue. I would like to take a look at [test-fest](https://github.com/sindresorhus/type-fest)
because they have some useful utility types that I may want to try in
this project like RequireAtLeastOne among others.

As a final note, not all routes are tested yet. This diff was just made
to fix the broken ones. I will still need to create tests for the others
which should not take long.