Database schema change with lots of changes
Description:
- Removed accounts table in favor of passwords table
- Moved email to users table
- Created passwords table
- Added indexes on various columns in various tables
- Updated all references to accountId in favor of userId
- Updated all queries to use userId instead of accountId
- Updated cached data to no longer user accountId
- Updated session type to no longer support accountId

Test Plan:
- Updated tests on @me route
- Ensure that other tests still pass

Notes:
Major update!

Ideally, this would have been multiple commits in a situation where I
have actual data, but since the tables are empty, I can expedite these
massive schema changes.

I decided that I wanted to have a separate passwords table rather than
the accounts table that I was using before. This allows me to have the
email stored in the users tables which shuold prevent situations where
the passwords table can be accidentally queried. There is now a 1:1
relationship between users and passwords which is ideal. This also has
the added benefit of allowing me to create users without passwords which
sounds bad, but can actually be really useful. For example, if a company
adds logs of users via a csv, they can without problem (although I don't
support that functionality just yet). I can hten check if the user has a
password set when they attempt to login, if they don't, I can send them
an email that takes them to a place where they can set the password. I
could also do what slack does and simply allow the user to login with
the email if they are okay with using email to sign on every time.

I also updated the flows for password update and account registration to
the new system. Thir required many changes on the queries folder, but
wasn't too bad.

Many routes needed to be updated as well, they now support the new
queries without issue. I would like to have some integration tests in
the future to ensure that things are actually running properly, but that
is for another time.

I've also gone ahead and updated the session data type which will now
prevent accountId from being stored on the session token. It won't
remove it on existing ones, but I can easily delete those as needed.

Beyond that, most of the routes that have changes are due to the way I
have to check for authorization in each of the routes. When I update my
authorization strategy, I will be remvoing those checks.

Finally, I updated the queries folders to be plural so that they match
the table names that they are connecting to.