Added missing route tests
Description:
- Added tests for /api/switches
- Added tests for /api/switches/:switchId

Test Plan:
- Added new tests
- Ensure that all tests pass

Issue:
[Add missing tests for routes](https://github.com/IzzyBeraja/Flags/issues/51)

Notes:
All routes now have tests, although they aren't in a good state. Since
queries does not have proper Error organization, it's not clear how any
of the queries can fail. This means that the API endpoints cannot
respond with a proper error and will in many cases simply spit out the
db error. This is fine for dev work, but a security issue in production.
I should focus on creating a way to properly handle errors soon.

On another note, a lot of tests would be simpler if route level
middleware was operational. Currently all of them need to check if the
userId field is non-null to do any actual work which is repetitive.