Using non-default exports on backend
Description:
- Modified exports to not use default and have specific exports

Notes:
I saw online that someone recommended not using default exports. I use
them pretty often in React, but can see the benefit on the backend. As
such, since it's an easy fix, I have updated everything to be an
explicit export. It actually makes some of the exports look nicer anyway