Adding simple authentication middleware on routes
Description:
- Adds middleware on routes that require authentication

Test Plan:
- I will need to add integration tests in the future to check for this

Notes:
I added a simple middleware that I am running on routes to check that
the user is currently logged in. I would like to integrate this with api
keys next.

I would like to change the structure of how I create routes in the
future to something like this:
```
type PostHandlingStuff = {
  Request: PostRequest;
  Response: PostResponse | ErrorType;
  Params: Params;
  Middleware: IsAuthenticated;
};
```

In its current incarnation, I was unable to have strict type safety on
my routes. I think due the nature of Express and underlying http
handling, it's rather complicated to have type narrowing. I saw that
Elysia tries to accomplish this through method chaining, but the end
result is kind of odd looking. I think I'm close to what I'd like to
accomplish, but the lack of perfect type safety kind of sucks. At least
this isn't exclusively an Express problem and seems to happen across all
other frameworks even ones in other languages.