Updated types for RouteHandlers
Description:
- Handlers are now typed in the format of HandlerShape
- AsyncHandler takes a HandlerShape and converts into request

Test Plan:
- Fix tests in next diff

Notes:
I updated the types to be in an easier format. Now, only the relevant
info for an endpoint is needed to properly type it. No need to import a
useless Params or Query type to satisfy the geneic Request.

The new structure is as follows:
```
type HandlerShape = {
  Request: unknown;
  Response: unknown;
  Error?: unknown;
  Params?: unknown;
  Query?: unknown;
  Middleware?: unknown;
};
```

Here is an example:
```
export type PatchHandler = {
  Params: Params;
  Response: PatchResponse;
  Request: PatchRequest;
  Error: ErrorType;
  Middleware: IsAuthenticated;
};
```

This makes it a lot clearer what the handler types are and makes adding
things like middleware significantly easier than before:
```
export type PatchHandler = AsyncHandler<
  Params,
  PatchResponse | ErrorType,
  PatchRequest,
  EmptyObject,
  IsAuthenticated
>;
```
Notice the EmtpyObject and Response union. If this route doesn't have a
Params field, you would have still needed to include it. Now you don't!