Updated tests to support new types and run faster
Description:
- Tests now support the new HandlerShape type and are simpler
- Removed various tests that check for authentication as they are no
longer needed
- Updated jest to use `isolatedModules: true` since tests were running
slowly

Test Plan:
- Ensure that all tests are now passing

Notes:
Fixed tests again. They now support the new syntax. Luckily it was a
drop in replacement. As well, I've gone ahead and removed type exports
for the handlers as they are not used for tests. I use Parameters
instead because it allows for the Request type to be properly generated.
I've also updated jest to include a new flag. This was done because
tests were taking over 7 seconds in total, now they have dropped to 1.5.
There is some type checking lost when run, but I can still see the
errors within the files.