Adds secret fetching from Google Cloud
Description:
- Secrets that are not available in .env will now be queried from google
secret manager

Notes:
This should make development a little easier on other machines, but I
don't think this will be the lasting solution. I think in an ideal world
I would be grabbing the secrets from an orchestration tool like
Kubernetes and then passing them to containers instead. The benefit of
this approach is that I can store these keys in a safe storage and
access them from machines that I trust.

There are some caveats:
- It's slow to query the keys from Google Secret Manager (GSM)
- I have limited reads and writes in the free tier
- Requires a Google authentication step that is different on local dev
than production environments

I do wonder if it makes sense to store these variables on my dev machine
on the first request to secret-manager. I would need to also store the
versions of the keys so that I can update older versions when the
secrets are updated.