Gather env variables for local development
Description:
- Runs a startup script that sets env variables on first startup

Issue:
[Learn about secret storage with google cloud](https://github.com/IzzyBeraja/Flags/issues/56)

Notes:
This will make development easier as the secrets are automatically
requested from GCP SM instead of having to store them elsewhere. There
are some issues with this however:

- env variables are stored in a persistent file that exists past the
lifetime of the server
- variables are not stored with version info so if they are old, they
need to be manually removed
- Not scalable to a larger group of developers

There are some other approaches that I can look into in the future:
- Create a development proxy so I hit faster cache than GCP
- Store environment variables in terminal env variables so they don't
persist
- Utilize tools like direnv to run scripts on directory open (eh...)

For now, this doesn't seem like a big issue as I'm only a singular dev,
but in the future I'd like a more concrete implementation that works for
more people. At least it's better than storing in notepad...