Adds support for api key storage
Description:
- Adds two new queries (createApiKey, getApiKey)
- Adds new route api-key with GET and POST methods
- Added new module that creates, encrypts, and decrypts api keys

Test Plan:
- Added tests for api-key route
- Added tests for apiKeyFunctions

Issue:
[Add API key support](https://github.com/IzzyBeraja/Flags/issues/43)

Notes:
This adds support for the creation of api keys. Currently they are
encrypted using AES, although there is no initilization vector (IV). The
main reason for this is because the api keys generated have sufficient
entropy and would only marginally benefit from the vector. Furthermore
the encrypted values are unique so there won't be any other keys in the
db that match. I've implemented retry logic to ensure that.

When the api keys are stored, they are encrypted within the database and
made visible to the client via response. When the keys are retrieved,
they are decrypted and shown to the client.

Currently there is no authentication around api keys just yet so nothing
can be done with them. There is also no limit on the amount of keys a
user can create. I may want to add a maximum in the future.